backend/requirements.txt
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.2
fastapi==0.115.6
google-auth==2.37.0
greenlet==3.1.1
griffe==1.5.4
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.8.2
jsonpath-python==1.0.6
logfire-api==3.1.0
mistralai==1.2.6
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
mysqlclient==2.2.7
openai==1.59.7
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic-ai==0.0.18
pydantic-ai-slim==0.0.18
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
redis==5.2.1
reportlab==4.2.5
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
